Extract widget
main.dartにウィジェットの定義も書かれていますが、
このままやっていくと、このファイルがとても大きくなってしまいます。

そこで、ホームページとして定義されているウィジェットの部分を
widgets/my_home_page.dartとして分割します。
